refactor(wave-4): decompose complex forms and wizard hooks
Decomposes monolithic hooks and components in Medicine, Stock, and Protocol modules to comply with strict linting rules (max-lines-per-function and complexity). Extracted submission logic into services and modularized UI into logical sections.
